LayoutTests/platform/glib/webgl/webgl-allow-shared-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_VALUE: bufferSubData: offset out of range
CONSOLE MESSAGE: WebGL: INVALID_ENUM: texImage2D: invalid texture type
CONSOLE MESSAGE: WebGL: INVALID_ENUM: texSubImage2D: invalid texture type
CONSOLE MESSAGE: WebGL: INVALID_ENUM: compressedTexImage2D: invalid format
CONSOLE MESSAGE: WebGL: INVALID_ENUM: compressedTexSubImage2D: invalid format
CONSOLE MESSAGE: WebGL: INVALID_ENUM: readPixels: invalid type


PASS WebGL AllowShared
FAIL WebGL2 AllowShared null is not an object (evaluating 'gl.createBuffer')
